Rename onboarding-screen folder to onboarding
