+++ src/onboarding/OnboardingFullScreenView.tsx
