+++ src/stacks-hierarchy/Root_LocationWhenInUsePermissionScreen.tsx
@@ -3,7 +3,7 @@ import React, {useCallback} from 'react';
 
 import LocationWhenInUsePermissionTexts from '@atb/translations/screens/LocationWhenInUsePermission';
 import {MyLocation} from '@atb/assets/svg/color/images';
-import {OnboardingScreenComponent} from '@atb/onboarding-screen';
+import {OnboardingScreenComponent} from '@atb/onboarding';
 import {useGeolocationState} from '@atb/GeolocationContext';
 import {useOnboardingNavigation} from '@atb/utils/use-onboarding-navigation';
